feat: switch to OSM within Events
